adding dry run

+++ api/management/commands/cleanposts.py
@@ -15,17 +15,28 @@ class Command(BaseCommand):
             help='No of days for which the posts are kept',
             default=31,
         )
+        parser.add_argument(
+            '--dry_run',
+            action='store_true',
+            dest='dry_run',
+            help='Dry run',
+            default=False,
+        )
 
     def handle(self, *args, **options):
         days = options['days']
+        dry_run = options['dry_run']
         post_eol = datetime.now() - timedelta(days=int(days))
         # delete all posts
         post_queryset = Post.objects.filter(last_modified_on__lt=post_eol)
         print("deleting {} posts".format(post_queryset.count()))
-        post_queryset.delete()
+        if not dry_run:
+            post_queryset.delete()
 
         # delete all media which are not referred in posts
         media_item_ids = [media_item.id for post in Post.objects.all() for media_item in post.media_items.all()]
         orphan_media = Media.objects.exclude(id__in=media_item_ids)
         print("deleting {} media items".format(orphan_media.count()))
-        orphan_media.delete()
+        # delete the media from cloudinary also
+        if not dry_run:
+            orphan_media.delete()
